updated mailchimp for wp, polylang, paypal, stripe payments, product gallery slider, and packaging slips plugins

+++ wp-content/plugins/pymntpl-paypal-woocommerce/src/Factories/ItemsFactory.php
@@ -24,9 +24,9 @@ class ItemsFactory extends AbstractFactory {
 			 * the per unit price since PayPal does their own unit_price * quantity.
 			 */
 			if ( $incl_tax ) {
-				$total = ( $cart_item['line_subtotal'] + $cart_item['line_subtotal_tax'] ) / (int) $cart_item['quantity'];
+				$total = ( $cart_item['line_subtotal'] + $cart_item['line_subtotal_tax'] ) / (float) $cart_item['quantity'];
 			} else {
-				$total = $cart_item['line_subtotal'] / (int) $cart_item['quantity'];
+				$total = $cart_item['line_subtotal'] / (float) $cart_item['quantity'];
 			}
 			$qty  = $cart_item['quantity'];
 			$name = $cart_item['data']->get_name();
